synq: split CLI runtime into per-command modules
Break syntaqlite-cli/src/runtime.rs (1.4k lines mixing dispatch,
source loading, schema/format/check config merging, and seven
subcommand bodies) into focused modules:

- cli.rs: clap arg types; each subcommand is a clap::Args struct
  wrapped in a Command tuple variant so dispatch is a trivial match
- util.rs: Source struct + load_sources, schema catalog construction,
  format/check config merging
- config.rs: adds ConfigMode + resolve alongside the existing TOML
  loader
- lib.rs: CliApp trait, run, and the dispatch match inline (no
  separate dispatch.rs)
- commands/{parse,fmt,validate,lineage,lsp,tokenize,mcp,codegen}.rs:
  one module per subcommand, each built around a runner struct whose
  methods replace the tangle of free helpers
